Discusion/src/pyproject.toml
[project]
name = "kabu-trading-src"
version = "0.1.0"
description = "株銘柄評価ツール"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk",
    "anyio",
    "pyyaml",
]

[dependency-groups]
dev = []
